first succesful R .Call

+++ src/poiss.c
@@ -3,7 +3,7 @@
 #include <gsl/gsl_randist.h>
 #include "poiss.h"
 
-void poiss_jump_times(double* t,double lambda,double n,unsigned long long seed){
+void poiss_jump_times(double* t,double lambda,size_t n,unsigned long long seed){
     // Generuje momenty n skoków procesu Poissona Nt
     // Zapisuje wynik do wektora t
