+++ src/poiss.h
@@ -2,8 +2,7 @@
 #define POISS.H
 
 
-
-void poiss_jump_times(double* t,double lambda,double n,unsigned long long seed);
+void poiss_jump_times(double* t,double lambda,size_t n,unsigned long long seed);
 
 
 #endif
